Add abstract version control handlers - (Task #539)
Refactor the handler class hierarchy:
- Add AVersionControlHandler
- Add AVersionControlUpdateHandler
- Add AVersionControlCommitHandler
- Add svn handlers
- Use new handlers in plugin.xml

---
Task #539: Refactor SVN backend
