+++ de.dlr.sc.virsat.team.ui/plugin.xml
@@ -23,7 +23,7 @@
    <extension
          point="org.eclipse.ui.handlers">
       <handler
-            class="de.dlr.sc.virsat.team.ui.action.svn.SVNCommitAction"
+            class="de.dlr.sc.virsat.team.ui.handler.svn.SvnCommitHandler"
             commandId="de.dlr.sc.virsat.team.ui.command.commit">
          <enabledWhen>
             <with
@@ -40,7 +40,7 @@
          </enabledWhen>
       </handler>
       <handler
-            class="de.dlr.sc.virsat.team.ui.action.svn.SVNUpdateAction"
+            class="de.dlr.sc.virsat.team.ui.handler.svn.SvnUpdateHandler"
             commandId="de.dlr.sc.virsat.team.ui.command.update">
          <activeWhen>
             <with
@@ -57,7 +57,7 @@
          </activeWhen>
       </handler>
  	  <handler
-            class="de.dlr.sc.virsat.team.ui.action.git.GitCommitHandler"
+            class="de.dlr.sc.virsat.team.ui.handler.git.GitCommitHandler"
             commandId="de.dlr.sc.virsat.team.ui.command.commit">
          <activeWhen>
             <with
@@ -73,7 +73,7 @@
          </activeWhen>
       </handler>
       <handler
-            class="de.dlr.sc.virsat.team.ui.action.git.GitUpdateHandler"
+            class="de.dlr.sc.virsat.team.ui.handler.git.GitUpdateHandler"
             commandId="de.dlr.sc.virsat.team.ui.command.update">
          <activeWhen>
             <with
